% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{add_metric}
\alias{add_metric}
\title{Apply a scalar metric in parallel using future_lapply}
\usage{
add_metric(sf_obj, calc_func, col_name, ncores = 1, quiet = FALSE, ...)
}
\arguments{
\item{sf_obj}{An sf object with multiple polygons.}

\item{calc_func}{A scalar function (calc_*) to apply to each polygon.}

\item{col_name}{Name of the new column to add.}

\item{ncores}{Number of cores for parallel processing. If \code{>1},
sets \code{future::plan("multisession", workers = ncores)}. 
If \code{NULL} or \code{<1}, it is set to 1 (no parallelization).}

\item{quiet}{Logical: if TRUE, suppress messages about core usage.}

\item{...}{Additional arguments passed to \code{calc_func}.}
}
\value{
The sf object with the new column added.
}
\description{
Apply a scalar metric in parallel using future_lapply
}
\keyword{internal}
