% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{check_sf}
\alias{check_sf}
\title{Check if object is an sf object and check CRS}
\usage{
check_sf(obj, arg_name = "sf_obj")
}
\arguments{
\item{obj}{Object to check.}

\item{arg_name}{Name of the argument (for error message).}
}
\value{
Invisibly returns TRUE if object is sf; otherwise stops with error.
}
\description{
Check if object is an sf object and check CRS
}
\keyword{internal}
