% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{check_valid_geometry}
\alias{check_valid_geometry}
\title{Check if all geometries in an sf object are valid}
\usage{
check_valid_geometry(sf_obj)
}
\arguments{
\item{sf_obj}{An sf object.}
}
\value{
Invisibly returns TRUE if all geometries are valid; otherwise stops with error.
}
\description{
Check if all geometries in an sf object are valid
}
\keyword{internal}
