% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{compactness_metrics}
\alias{compactness_metrics}
\title{Compactness Metrics (redistmetrics style)}
\usage{
compactness_metrics(sf_obj, ncores = 1, ...)
}
\arguments{
\item{sf_obj}{An sf object.}

\item{ncores}{Number of cores for each metric. Default is 1 (sequential).}

\item{...}{Additional arguments passed to each metric function.}
}
\value{
The sf object with compactness columns added.
}
\description{
Computes a subset of metrics focused on compactness as defined in the
\code{redistmetrics} package.
}
\references{
Based on the metrics described in the redistmetrics package.
}
