% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{disable_s2_if_geographic}
\alias{disable_s2_if_geographic}
\title{Disable s2 for geographic CRS and return previous state}
\usage{
disable_s2_if_geographic(sf_obj, func_name = "function")
}
\arguments{
\item{sf_obj}{An sf object.}

\item{func_name}{Name of the calling function (for the message).}
}
\value{
A logical value: the previous state of s2 (TRUE if s2 was enabled), or NULL if no change.
}
\description{
If the sf object has a geographic CRS and s2 is currently enabled,
this function disables s2 and informs the user. It returns the previous
state of s2 so it can be restored later.
}
\keyword{internal}
