% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{fractal_dimension}
\alias{fractal_dimension}
\title{Fractal Dimension}
\usage{
fractal_dimension(sf_obj, ncores = 1, quiet = FALSE)
}
\arguments{
\item{sf_obj}{An sf object containing polygon geometries.}

\item{ncores}{Number of cores for parallel computation. Default is 1 (sequential).}

\item{quiet}{Logical: suppress messages about core usage.}
}
\value{
sf object with column `fractal_dimension`.
}
\description{
Approximate fractal dimension: \eqn{2 * \log(P) / \log(A)}.
}
\references{
Mandelbrot, B. B. (1983). *The Fractal Geometry of Nature*. W. H. Freeman.
Also used in FRAGSTATS.
}
