% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{proximity}
\alias{proximity}
\title{Proximity Index}
\usage{
proximity(sf_obj, ncores = 1, n_samples = 1000, quiet = FALSE)
}
\arguments{
\item{sf_obj}{An sf object containing polygon geometries.}

\item{ncores}{Number of cores for parallel computation. Default is 1 (sequential).}

\item{n_samples}{Number of random points to sample inside the polygon (default: 1000).}

\item{quiet}{Logical: suppress messages about core usage.}
}
\value{
The sf object with an additional column `proximity`.
}
\description{
Ratio of the mean distance from the centroid in the equal‑area circle to
that in the polygon. Values close to 1 indicate that the polygon's internal
distances are similar to a circle.
}
\references{
McGarigal, K., Cushman, S. A., & Ene, E. (2012). FRAGSTATS v4.
}
