% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{schwartzberg}
\alias{schwartzberg}
\title{Schwartzberg Compactness}
\usage{
schwartzberg(sf_obj, ncores = 1, quiet = FALSE)
}
\arguments{
\item{sf_obj}{An sf object containing polygon geometries.}

\item{ncores}{Number of cores for parallel computation. Default is 1 (sequential).}

\item{quiet}{Logical: suppress messages about core usage.}
}
\value{
sf object with column `schwartzberg`.
}
\description{
Computes \eqn{1 / (P / (2\pi \sqrt{A/\pi}))}.
}
\references{
Schwartzberg, J. E. (1966). Reapportionment, gerrymanders, and the
1964 elections. *The Annals of the American Academy of Political and Social Science*, 363(1), 54-68.
}
