% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{shape_index}
\alias{shape_index}
\title{Shape Index}
\usage{
shape_index(sf_obj, ncores = 1, quiet = FALSE)
}
\arguments{
\item{sf_obj}{An sf object containing polygon geometries.}

\item{ncores}{Number of cores for parallel computation. Default is 1 (sequential).}

\item{quiet}{Logical: suppress messages about core usage.}
}
\value{
sf object with column `shape_index`.
}
\description{
\eqn{P / (2 * \sqrt{\pi A})}. Minimum value is 1 for a circle.
}
\references{
McGarigal, K., Cushman, S. A., & Ene, E. (2012). FRAGSTATS v4.
}
