% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{shape_metrics}
\alias{shape_metrics}
\title{Shape Metrics (vectormetrics patch style)}
\usage{
shape_metrics(sf_obj, ncores = 1, ...)
}
\arguments{
\item{sf_obj}{An sf object.}

\item{ncores}{Number of cores for each metric. Default is 1 (sequential).}

\item{...}{Additional arguments passed to each metric function.}
}
\value{
The sf object with shape metrics columns added.
}
\description{
Computes a subset of metrics focused on shape, inspired by the patch metrics
in the \code{vectormetrics} package.
}
\references{
Inspired by the patch metrics described in the vectormetrics package.
}
