% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{skew}
\alias{skew}
\title{Skew Compactness}
\usage{
skew(sf_obj, ncores = 1, quiet = FALSE)
}
\arguments{
\item{sf_obj}{An sf object containing polygon geometries.}

\item{ncores}{Number of cores for parallel computation. Default is 1 (sequential).}

\item{quiet}{Logical: suppress messages about core usage.}
}
\value{
sf object with column `skew`.
}
\description{
Ratio of area of maximum inscribed circle to area of minimum circumscribing circle.
}
\references{
McGarigal, K., Cushman, S. A., & Ene, E. (2012). FRAGSTATS v4.
}
