% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_metrics.R
\name{check_polygon_geometry}
\alias{check_polygon_geometry}
\title{Check if all geometries in an sf object are polygons}
\usage{
check_polygon_geometry(sf_obj)
}
\arguments{
\item{sf_obj}{An sf object.}
}
\value{
Invisibly returns TRUE if all geometries are polygons; otherwise stops with error.
}
\description{
Check if all geometries in an sf object are polygons
}
\keyword{internal}
